Move all layer tree operations to a ScrollTree struct
This will make it easier to introduce the idea of reference frames
into WebRender. This should not change behavior.
